Fix various bugs.
First bug: in find_endpoint_bound, there was a line like list_1 = list_2,
where we wanted to update list_1 but leave list_2 static. The solution was to
use copy.deepcopy() (since they are lists of lists.

Second bug: If we want to move out the middle of some endpoints, and the two
endpoints were identical, there would be a division by zero. The solution was
to not do that: whenever we wanted to do that operation, we just wanted to
move the middle out maximally, so in this case we could just return one of
the endpoints without actually having to do any of the computation.